requirement.txt
absl-py==1.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
av==8.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
cachetools==5.3.0
cattrs==23.1.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
comm==0.1.2
compressai==1.2.3
contourpy==1.0.7
coremltools==7.0b1
crypto==1.4.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docopt==0.6.2
duckdb==0.7.2.dev457
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.12.0
fonttools==4.38.0
fqdn==1.5.1
future==0.18.3
gdown==4.6.4
Glymur==0.12.2
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
imageio==2.25.0
importlib-metadata==6.6.0
importlib-resources==5.10.2
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdv @ git+http://github.com/axiros/terminal_markdown_viewer@435b6885d0237fe10a9f894375b10c1103bfaae0
mistune==2.0.4
motion-vector-extractor==1.0.6
mpmath==1.3.0
Naked==0.1.32
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.5.5.64
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.4.29
pipreqs==0.4.11
pkgconfig==1.5.5
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
plotille==4.0.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch-msssim==0.2.1
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
range-coder==1.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.19.3
scipy==1.8.1
Send2Trash==1.8.0
shellescape==3.8.1
six==1.16.0
snakeviz==2.1.1
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminado==0.17.1
tifffile==2023.1.23.1
tinycss2==1.2.1
torch==1.13.1
torchac==0.9.3
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.3.2
tqdm==4.62.3
traitlets==5.8.1
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.15.0
